Regression Start and Testing
I started work on the regression problem. Still need to apply it to the other models.

Also added the beginnings to the model-testing notebook. Will need to export all models in main.ipynb and then add the procedures to the testing file. Just define the models then import the weights from the exported pth file. It's been done for Regression CNN and can easily be done for the others. Then add a readme for explaining how to import the models.